+++ packages/bundler/CHANGELOG.md
@@ -1,5 +1,71 @@
 # @0xsodium/bundler
 
+## 1.0.0
+
+### Major Changes
+
+- release
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@1.0.0
+
+## 0.1.0
+
+### Minor Changes
+
+- fix versions
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.1.0
+
+## 0.0.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.0.0
+
+## 0.0.0-20230218225937
+
+### Minor Changes
+
+- Fix app provider
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.0.0-20230218225937
+
+## 0.0.0-20230218190034
+
+### Minor Changes
+
+- Support app provider transports
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.0.0-20230218190034
+
+## 0.0.0-20230218181618
+
+### Minor Changes
+
+- Support app provider transports
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@0.0.0-20230218181618
+
 ## 0.21.0
 
 ### Minor Changes